+++ workers/ohsome_quality_analyst/oqt.py
@@ -34,10 +34,10 @@ from ohsome_quality_analyst.utils.exceptions import (
     EmptyRecordError,
     SizeRestrictionError,
 )
-from ohsome_quality_analyst.utils.helper import (
+from ohsome_quality_analyst.utils.helper import loads_geojson, name_to_class
+from ohsome_quality_analyst.utils.helper_asyncio import (
+    filter_exceptions,
     gather_with_semaphore,
-    loads_geojson,
-    name_to_class,
 )
 
 
@@ -76,7 +76,9 @@ async def _(
             await check_area_size(feature.geometry)
         tasks.append(create_indicator(parameters.copy(update={"bpolys": feature})))
     indicators = await gather_with_semaphore(tasks)
-    features = [i.as_feature(flatten=parameters.flatten) for i in indicators]
+    features = [
+        i.as_feature(parameters.flatten, parameters.include_data) for i in indicators
+    ]
     if len(features) == 1:
         return features[0]
     else:
@@ -91,7 +93,7 @@ async def _(
 ) -> Feature:
     """Create an indicator as GeoJSON object."""
     indicator = await create_indicator(parameters, force)
-    return indicator.as_feature(flatten=parameters.flatten)
+    return indicator.as_feature(parameters.flatten, parameters.include_data)
 
 
 async def create_report_as_geojson(
@@ -122,14 +124,16 @@ async def create_report_as_geojson(
                 parameters.copy(update={"bpolys": feature}),
                 force,
             )
-            features.append(report.as_feature(flatten=parameters.flatten))
+            features.append(
+                report.as_feature(parameters.flatten, parameters.include_data)
+            )
         if len(features) == 1:
             return features[0]
         else:
             return FeatureCollection(features=features)
     elif isinstance(parameters, ReportDatabase):
         report = await create_report(parameters, force)
-        return report.as_feature(flatten=parameters.flatten)
+        return report.as_feature(parameters.flatten, parameters.include_data)
     else:
         raise ValueError("Unexpected parameters: " + str(parameters))
 
@@ -387,7 +391,12 @@ async def create_all_indicators(
                     force=force,
                 )
             )
-    await gather_with_semaphore(tasks)
+    # Do no raise exceptions. Filter out exceptions from result list and log them.
+    results = await gather_with_semaphore(tasks, return_exceptions=True)
+    exceptions = filter_exceptions(results)
+    for exception in exceptions:
+        message = getattr(exception, "message", repr(exception))
+        logging.warning("Ignoring error: {0}".format(message))
 
 
 async def check_area_size(geom: Union[Polygon, MultiPolygon]):