+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Current Main
 
+### Breaking Changes
+
+- Make inclusion of indicator data in response optional ([#370])
+
+### How to Upgrade
+
+- To continue to retrieve additional data of an Indicator or Report provided in an API response, you need to set the parameter `include_data` to `True` when making requests to the API ([#370])
+
+[#370]: https://github.com/GIScience/ohsome-quality-analyst/pull/370
+
+## 0.10.0
+
 ### Bug Fixes
 
 - Add missing error handling of `RasterDatasetUndefinedError` and `RasterDatasetNotFoundError` to the API ([#298])
@@ -11,8 +23,9 @@
 ### New Features
 
 - Add support for `groupBy/boundary` queries to the ohsome API client ([#272])
-- Add `flatten` parameter to API request. Make flatten of GeoJSON properties of Indicators and Reports optional. ([#303])
+- Add `flatten` parameter to API request. Make flatten of GeoJSON properties of Indicators and Reports optional ([#303])
 - Make calculation of an Indicator for a FeatureCollection or for a Report asynchronous ([#307])
+- Add new Indicator which predicts the building area of the AOI using a trained Random Forest Regressor ([#265])
 
 ### Other Changes
 
@@ -23,6 +36,7 @@
 - Add hex-cells at zoom level 12 for Africa to the database ([#314])
 - Disable size limit on input AOI if OSM data is provided through a request with a custom Layer object ([#330])
 
+[#265]: https://github.com/GIScience/ohsome-quality-analyst/pull/265
 [#272]: https://github.com/GIScience/ohsome-quality-analyst/pull/272
 [#298]: https://github.com/GIScience/ohsome-quality-analyst/pull/298
 [#299]: https://github.com/GIScience/ohsome-quality-analyst/pull/299